refactor: 일간분석, 월간분석 데이터 매핑함수 분리 (#27)
