+++ src/components/page/strategy-detail/tabmenu/MonthlyAnalysis.tsx
@@ -9,17 +9,7 @@ import Pagination from '@/components/common/Pagination';
 import useFetchMonthlyAnalysis from '@/hooks/queries/useFetchMonthlyAnalysis';
 import usePagination from '@/hooks/usePagination';
 import { UserRole } from '@/types/route';
-
-export interface MonthlyDataProps {
-  strategyMonthlyDataId: number;
-  analysisMonth: string;
-  monthlyAveragePrincipal: number;
-  monthlyDepWdAmount: number;
-  monthlyPl: number;
-  monthlyReturn: number;
-  monthlyCumulativePl: number;
-  monthlyCumulativeReturn: number;
-}
+import { mapToMonthlyAnalysisData } from '@/utils/mappers';
 
 const MonthlyAnalysis = ({ attributes, strategyId, role }: AnalysisProps) => {
   const { pagination, setPage } = usePagination(1, 8);
@@ -32,16 +22,7 @@ const MonthlyAnalysis = ({ attributes, strategyId, role }: AnalysisProps) => {
 
   const normalizedData = useMemo(() => {
     if (!monthlyAnalysis) return [];
-    return monthlyAnalysis.map((data: MonthlyDataProps) => ({
-      dataId: data.strategyMonthlyDataId,
-      date: data.analysisMonth,
-      principal: data.monthlyAveragePrincipal,
-      dep_wd_price: data.monthlyDepWdAmount,
-      profit_loss: data.monthlyPl,
-      pl_rate: data.monthlyReturn,
-      cumulative_profit_loss: data.monthlyCumulativePl,
-      cumulative_profit_loss_rate: data.monthlyCumulativeReturn,
-    }));
+    return monthlyAnalysis.map(mapToMonthlyAnalysisData);
   }, [monthlyAnalysis]);
 
   if (isLoading) {
