+++ src/types/strategyDetail.ts
@@ -29,6 +29,7 @@ export interface InputDailyAnalysisProps {
   payload: DailyAnalysisProps[];
 }
 
+//일간분석 테이블 데이터
 export interface AnalysisDataProps {
   dailyStrategicStatisticsId: number;
   inputDate: string;
@@ -38,6 +39,18 @@ export interface AnalysisDataProps {
   dailyPlRate: number;
   cumulativeProfitLoss: number;
   cumulativeProfitLossRate: number;
+}
+
+//월간분석 테이블 데이터
+export interface MonthlyDataProps {
+  strategyMonthlyDataId: number;
+  analysisMonth: string;
+  monthlyAveragePrincipal: number;
+  monthlyDepWdAmount: number;
+  monthlyPl: number;
+  monthlyReturn: number;
+  monthlyCumulativePl: number;
+  monthlyCumulativeReturn: number;
 }
 
 export interface StatisticsProps {
